% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covsep.R
\docType{data}
\name{C1}
\alias{C1}
\title{A covariance matrix}
\format{\preformatted{ num [1:32, 1:32] 0.1036 0.0726 0.0654 0.063 0.0891 ...
}}
\usage{
C1
}
\description{
Marginal covariance matrix C1 used for simulations in the paper \url{http://arxiv.org/abs/1505.02023}
}
\section{Details}{

This is a 32x32 real-valued covariance matrix.
}
\examples{
data(C1)
str(C1)
}
\references{
\cite{John A. D. Aston, Davide Pigoli, Shahin Tavakoli, "Tests
for separability in nonparametric covariance operators of random surfaces",
2015, under revision}, \url{http://arxiv.org/abs/1505.02023}
}
\keyword{datasets}

