% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covsep.R
\docType{data}
\name{C2}
\alias{C2}
\title{A covariance matrix}
\format{An object of class \code{matrix} with 7 rows and 7 columns.}
\usage{
C2
}
\description{
Marginal covariance matrix C2 used for simulations in paper
\url{http://arxiv.org/abs/1505.02023}
}
\section{Details}{
 This is a 7x7 real-valued covariance matrix.
}

\examples{
data(C2)
str(C2)

}
\references{
\cite{Aston, John A. D.; Pigoli, Davide; Tavakoli, Shahin. Tests for
separability in nonparametric covariance operators of random surfaces. Ann.
Statist. 45 (2017), no. 4, 1431--1461. doi:10.1214/16-AOS1495.
https://projecteuclid.org/euclid.aos/1498636862}
}
\keyword{datasets}
