% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covsep.R
\docType{package}
\name{covsep}
\alias{covsep}
\alias{covsep-package}
\title{covsep: tests for determining if the covariance structure of
2-dimensional data is separable}
\description{
Functions for testing if the covariance structure of 2-dimensional data (e.g.
samples of surfaces X_i = X_i(s,t)) is separable, i.e. if cov(X) = C_1 x C_2.
A complete descriptions of the implemented tests can be found in the paper
arXiv:1505.02023.
}
\section{Main functions}{

The main functions are \itemize{
\item \code{\link{clt_test}},
\item \code{\link{gaussian_bootstrap_test}},
\item \code{\link{empirical_bootstrap_test}},
\item \code{\link{HS_gaussian_bootstrap_test}},
\item \code{\link{HS_empirical_bootstrap_test}}
}
}

