% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covsep.R
\name{difference_fullcov}
\alias{difference_fullcov}
\title{compute the difference between the full sample covariance and its separable
approximation}
\usage{
difference_fullcov(Data)
}
\arguments{
\item{Data}{a (non-empty) \code{N x d1 x d2} array of data values. The first
direction indices the \eqn{N} observations, each consisting of a \code{d1 x d2}
discretization of the surface, so that \code{Data[i,,]} corresponds to the
i-th observed surface.}
}
\value{
A \code{d1 x d2 x d1 x d2} array, where \code{d1 = nrow(Data)} and
\code{d2 = ncol(Data)}.
}
\description{
compute the difference between the full sample covariance and its separable
approximation
}
\section{Details}{

This is an internal function.
}

