% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covsep.R
\name{rmtnorm}
\alias{rmtnorm}
\title{Generate a sample from a Matrix Gaussian distribution}
\usage{
rmtnorm(N, C1, C2, M = matrix(0, nrow(C1), nrow(C2)))
}
\arguments{
\item{N}{sample size}

\item{C1}{row covariance}

\item{C2}{column covariance}

\item{M}{mean matrix}
}
\value{
A \code{N x dim(C1)[1] x dim(C2)[1]} array containing the
generated data
}
\description{
Generate a sample from a Matrix Gaussian distribution
}
\examples{
Data = rmtnorm(30, C1, C2)
}

