% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covsep.R
\docType{data}
\name{SurfacesData}
\alias{SurfacesData}
\title{A data set of surfaces}
\format{\preformatted{ num [1:50, 1:32, 1:7] -0.229 0.64 0.373 0.346 -0.402 ...
}}
\usage{
SurfacesData
}
\description{
Dataset of 50 surfaces simulated from a Gaussian process with a separable covariance structure. \code{SurfacesData[i,,]} corresponds to the i-th surface observed on a 32x7 uniform grid.
}
\section{Details}{

 This is a \code{50 x 32 x 7} array.
}
\examples{
data(SurfacesData)
image(SurfacesData[1,,]) # color image of the first surface in the dataset
}
\keyword{datasets}

