% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covsep.R
\docType{data}
\name{C2}
\alias{C2}
\title{A covariance matrix}
\format{\preformatted{ num [1:7, 1:7] 1.598 1.604 1.384 1.167 0.909 ...
}}
\usage{
C2
}
\description{
Marginal covariance matrix C2 used for simulations in paper
\url{http://arxiv.org/abs/1505.02023}
}
\section{Details}{
 This is a 7x7 real-valued covariance matrix.
}
\examples{
data(C2)
str(C2)
}
\references{
\cite{John A. D. Aston, Davide Pigoli, Shahin Tavakoli, "Tests
for separability in nonparametric covariance operators of random surfaces",
2015, under revision}, \url{http://arxiv.org/abs/1505.02023}
}
\keyword{datasets}

