% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covsep.R
\docType{data}
\name{SurfacesData}
\alias{SurfacesData}
\title{A data set of surfaces}
\format{An object of class \code{array} of dimension 50 x 32 x 7.}
\usage{
SurfacesData
}
\description{
Dataset of 50 surfaces simulated from a Gaussian process with a separable covariance structure. \code{SurfacesData[i,,]} corresponds to the i-th surface observed on a 32x7 uniform grid.
}
\section{Details}{

 This is a \code{50 x 32 x 7} array.
}

\examples{
data(SurfacesData)
image(SurfacesData[1,,]) # color image of the first surface in the dataset
}
\keyword{datasets}
